(grpcxx) move libuv dependent code to its own namespace
