+++ lib/grpcxx/CMakeLists.txt
@@ -1,35 +1,48 @@
 add_library(grpcxx)
 target_sources(grpcxx
 	PRIVATE
-		conn.cpp
 		context.cpp
 		message.cpp
-		reader.cpp
 		request.cpp
-		scheduler.cpp
-		server.cpp
-		writer.cpp
+		server_base.cpp
 	PUBLIC
 		FILE_SET headers TYPE HEADERS
 		FILES
 			context.h
 			fixed_string.h
 			rpc.h
-			scheduler.h
 			server.h
+			server_base.h
 			service.h
 			status.h
-			task.h
 	PRIVATE
 		FILE_SET private_headers TYPE HEADERS
 		FILES
-			conn.h
-			coroutine.h
 			message.h
-			reader.h
 			request.h
 			response.h
-			writer.h
+)
+
+target_sources(grpcxx
+	PRIVATE
+		uv/conn.cpp
+		uv/reader.cpp
+		uv/scheduler.cpp
+		uv/server.cpp
+		uv/writer.cpp
+	PUBLIC
+	FILE_SET headers TYPE HEADERS
+	FILES
+		uv/server.h
+		uv/scheduler.h
+		uv/task.h
+	PRIVATE
+	FILE_SET private_headers TYPE HEADERS
+	FILES
+		uv/conn.h
+		uv/coroutine.h
+		uv/reader.h
+		uv/writer.h
 )
 
 # h2 sources
