+++ lib/grpcxx/server_base.cpp
@@ -1,0 +1,31 @@
+#include "server_base.h"
+
+#include "request.h"
+#include "response.h"
+
+namespace grpcxx {
+namespace detail {
+response server_base::process(const request &req) const noexcept {
+	if (!req) {
+		return {req.id(), status::code_t::invalid_argument};
+	}
+
+	auto it = _services.find(req.service());
+	if (it == _services.end()) {
+		return {req.id(), status::code_t::not_found};
+	}
+
+	context          ctx(req);
+	detail::response resp(req.id());
+	try {
+		auto r = it->second(ctx, req.method(), req.data());
+		resp.status(std::move(r.first));
+		resp.data(std::move(r.second));
+	} catch (std::exception &e) {
+		return {req.id(), status::code_t::internal};
+	}
+
+	return resp;
+}
+} // namespace detail
+} // namespace grpcxx
