+++ lib/grpcxx/uv/conn.h
@@ -8,22 +8,22 @@
 
 #include <uv.h>
 
-#include "h2/session.h"
-
+#include "../h2/session.h"
+#include "../request.h"
+#include "../response.h"
 #include "reader.h"
-#include "request.h"
-#include "response.h"
 #include "task.h"
 #include "writer.h"
 
 namespace grpcxx {
+namespace uv {
 namespace detail {
 class conn {
 public:
 	using buffer_t   = std::string;
 	using handle_t   = std::shared_ptr<uv_tcp_t>;
-	using requests_t = std::forward_list<request>;
-	using streams_t  = std::unordered_map<int32_t, request>;
+	using requests_t = std::forward_list<::grpcxx::detail::request>;
+	using streams_t  = std::unordered_map<int32_t, ::grpcxx::detail::request>;
 
 	conn(const conn &) = delete;
 	conn(uv_stream_t *stream);
@@ -33,7 +33,7 @@ public:
 	requests_t   read(std::string_view bytes);
 	class reader reader() const noexcept;
 
-	void write(response resp) noexcept;
+	void write(grpcxx::detail::response resp) noexcept;
 
 private:
 	struct deleter {
@@ -53,4 +53,5 @@ private:
 	streams_t   _streams;
 };
 } // namespace detail
+} // namespace uv
 } // namespace grpcxx
